nist/requirements.txt
absl-py==1.0.0
adjusttext==0.7.3
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.8
anyio==3.5.0
argon2-cffi-bindings==21.2.0
argon2-cffi==21.3.0
astroid==2.9.3
async-timeout==4.0.2
attrs==20.3.0
autopep8==1.7.0
babel==2.9.1
backcall==0.2.0
bleach==4.1.0
blessings==1.7
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
colorlog==6.6.0
commonmark==0.9.1
configparser==5.2.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
docker-pycreds==0.4.0
einops==0.3.0
entrypoints==0.3
filelock==3.4.2
flake8==5.0.4
fonttools==4.29.0
frozenlist==1.3.0
fsspec==2022.2.0
future==0.18.2
geomloss==0.2.5
gitdb==4.0.9
gitpython==3.1.26
google-auth-oauthlib==0.4.6
google-auth==2.5.0
gpustat==0.6.0
grpcio==1.43.0
h5py==3.6.0
htmlmin==0.1.12
hydra-colorlog==1.1.0
hydra-core==1.1.1
hydra-joblib-launcher==1.1.5
identify==2.4.6
idna==3.3
imagehash==4.2.1
imageio==2.14.1
importlib-metadata==4.10.1
importlib-resources==5.4.0
iniconfig==1.1.1
ipdb==0.13.9
ipykernel==6.7.0
ipython-genutils==0.2.0
ipython==7.31.1
ipywidgets==7.6.5
isort==5.10.1
jammy==0.0.6
jedi==0.18.1
jinja2==3.0.3
joblib==1.0.1
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.1
jupyter-server==1.13.4
jupyter==1.0.0
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.0.2
jupyterlab==3.2.8
keopscore==2.1
kiwisolver==1.3.2
# knn-cuda==0.2
kornia==0.5.11
lazy-object-proxy==1.7.1
line-profiler==3.4.0
lmdb==1.3.0
loguru==0.5.3
markdown==3.3.6
markupsafe==2.0.1
matplotlib-inline==0.1.3
matplotlib==3.5.1
mccabe==0.7.0
missingno==0.5.1
mistune==0.8.4
msgpack-numpy==0.4.7.1
msgpack==1.0.3
multidict==6.0.2
multimethod==1.7
munkres==1.1.4
nbclassic==0.3.5
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
ninja==1.10.2.3
nodeenv==1.6.0
notebook==6.4.8
numpy==1.21.5
nvidia-ml-py3==7.352.0
oauthlib==3.2.0
omegaconf==2.1.1
ordered-set==4.1.0
# otdd-map==0.0.1
packaging==21.3
pandas-profiling==3.1.0
pandas==1.1.5
pandocfilters==1.5.0
parse==1.19.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
phik==0.12.0
pickleshare==0.7.5
pillow==9.0.0
pip==22.2.2
platformdirs==2.4.1
pluggy==1.0.0
pot==0.8.2
pre-commit==2.20.0
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.26
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pudb==2022.1
py==1.11.0
pyarrow==6.0.1
pyasn1-modules==0.2.8
pyasn1==0.4.8
pybind11==2.10.0
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.9.0
pydeprecate==0.3.1
pyflakes==2.5.0
pygments==2.11.2
pykeops==2.1
pylint==2.12.2
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
python-memcached==1.59
python-ternary==1.0.8
pytorch-lightning==1.5.2
pytz==2021.3
pywavelets==1.2.0
pyyaml==6.0
pyzmq==22.3.0
qtconsole==5.2.2
qtpy==2.0.1
requests-oauthlib==1.3.1
requests==2.27.1
rich==10.16.2
rsa==4.8
scikit-image==0.18.3
scikit-learn==0.24.2
scipy==1.7.3
seaborn==0.11.2
send2trash==1.8.0
sentry-sdk==1.5.4
setuptools-scm==6.4.2
setuptools==59.5.0
sh==1.14.2
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
subprocess32==3.5.4
tangled-up-in-unicode==0.1.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboard==2.10.0
terminado==0.13.1
testpath==0.5.0
threadpoolctl==3.0.0
tifffile==2021.11.2
toml==0.10.2
tomli==2.0.0
torch-ema==0.3
torch-fidelity==0.3.0
# torch==1.11.0+cu113
# torchaudio==0.11.0+cu113
torchinfo==1.6.3
torchmetrics==0.9.3
# torchvision==0.12.0+cu113
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing-extensions==4.0.1
ubelt==0.10.2
urllib3==1.25.11
urwid-readline==0.13
urwid==2.1.2
virtualenv==20.13.0
visions==0.7.4
wandb==0.10.33
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
werkzeug==2.0.2
wheel==0.37.1
widgetsnbextension==3.5.2
wrapt==1.13.3
yarl==1.7.2
zipp==3.7.0